\name{gRumble-package}
\alias{gRumble-package}
\alias{gRumble}
\docType{package}
\title{
Orient IMU data
~~ gRumble ~~
}
\description{
A package designed for handleing data from animal borne IMU data loggers. The package is mainly focused on correcting orientation of tags of unknown positions, such as data loggers that were fed to sharks or to tags that have shifted orientation once they were placed on animals.
}
\details{
\tabular{ll}{
Package: \tab gRumble\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-05-10\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Connor F. White
Jerry Moxley
Salvador J. Jorgensen


Maintainer: Connor F. White <cwhite@mbayaq.org>

}
\references{
~~ Literature or other references for background information ~~
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
